+++ docs-app/publish.js
@@ -1,6 +1,7 @@
 var ghPages = require("gh-pages");
 
-var packageUrl = "https://github.com/Shmew/Feliz.MaterialUI.git";
+//var packageUrl = "https://github.com/Shmew/Feliz.MaterialUI.git";
+var packageUrl = "https://github.com/ArtemyB/Feliz.MaterialUI.git";
 
 console.log("Publishing to ", packageUrl);